eval3d-pipeline/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "eval3d-pipeline"
version = "0.1.0"
description = "CLI wrapper around Eval3D for evaluating 3D assets (.obj, .glb, .ply) with uv."
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
  "typer>=0.9,<0.10",
  "click>=8.0,<8.2",
  "rich>=13,<14",
  "python-dotenv>=1,<2",
  "pydantic>=2,<3",
  "pydantic-settings>=2,<3",
]

[project.optional-dependencies]
render = [
  "trimesh>=4.0",
  "pyrender>=0.1.45",
  "opencv-python>=4.8",
  "PyOpenGL>=3.1",
  "numpy>=1.24",
  "Pillow>=10.0",
  "torch>=2.0",
]

[project.scripts]
eval3d-pipeline = "eval3d_pipeline.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/eval3d_pipeline"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv]
dev-dependencies = []

